Add more GCC options (and annotations)
This caused a new warning in last_api_error.cpp, which has been
silenced.

+++ breath/diagnostics/brt/dep/syst/windows/last_api_error.cpp
@@ -34,7 +34,9 @@ last_api_error::last_api_error( char const * p ) noexcept
     DWORD const         dw = ::FormatMessageA(
         FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS,
         nullptr,
-        m_last_error,
+        // this cast silences a -Wconversion warning under Cygwin64
+        // (which typedefs DWORD to unsigned int)
+        static_cast< DWORD >( m_last_error ),
         0,
         m_message + offset,
         static_cast< DWORD >( sizeof m_message - offset ),
